+++ RELEASE_NOTES.md
@@ -11,6 +11,11 @@
 
 ## Identity
 
+- Flags
+  - split of flags in 2 lists: a short-list, automatically imported via CSS, ready to use simply by using the CSS class, as it was before. about 10 countries flags. For both ratios square + 4/3
+  - list of flags in short-list : ["dk", "ee", "eu", "fi", "gb", "is", "lt", "lv", "no", "se", "us"]
+  - an extended list (the other 240 countries usually never used by any projects in here) that can be imported via your own custom CSS. Examples are added to the documentation.
+
 ## PayEx side UI changes
 
 ## Technical changes